[calendar] Use project calendar status by default, add "quiet" mode
The calendar update tool needed to be run manually, partly because we didn't
feel like automating it, and partly because there was no way to specify the
calendar entry status to use (draft, tentative or confirmed).

The script now looks for a `calendar` property in the project's metadata
(to be set in the project's description) and uses that status to determine
whether it should:
- do nothing (no property or property set to `no`), which is typically useful
during the first phase, when sessions get collected.
- create a `draft` calendar entry, which should only be useful on a temporary
basis to make sure that things work as intended.
- create a `tentative` calendar entry, for use once a first grid has been
announced.
- create a `confirmed` calendar entry, for use once the final grid has been
announced.

The script now also features a `quiet`, which makes it exit quietly without
throwing an error even if the session contains errors that need fixing.

Changes make it possible to automate calendar updates, as envisioned in #68.