Adjust validation logic for plenary sessions
See #74.

Validation will no longer report errors for plenary sessions scheduled at the
same time and in the same room, since that is to be expected.

Validation will now report a number of scheduling errors linked to the plenary
status of a session: too many sessions in a plenary, non empty list of
conflicting sessions, etc.