Handle plenary sessions (#75)
See #74.

Validation will no longer report errors for plenary sessions scheduled at the same time and in the same room, since that is to be expected. Validation will now report a number of scheduling errors linked to the plenary status of a session: too many sessions in a plenary, non empty list of conflicting sessions, etc. A warning is also returned when a session is scheduled in parallel with a plenary session (that may be needed but that's not normal).

The scheduling algorithm was adjusted to schedule sessions in plenary first and schedule multiple sessions in the same plenary slot.

The code has not been tested so probably buggy, but that should be easy to fix once we need it.

Note calendar management is still missing.