Fix project's info retrieval when project is empty
The list of labels that can be used is not retrieved from the project in
practice, but rather from the underlying issues repository. But that repository
is only know once a first issue has been linked to the project. The code now
returns an empty list of labels when there are no sessions linked to the
project yet.