Allow order of plenary sessions to be specified in notes
See discussion in #74.

By default, plenary sessions are ordered by issue number when they are
scheduled in the same plenary meeting. To override the order, a breakout
organizer may now add something like `order:1` to the Note field of the issue
in the underlying GitHub project.

Sessions with an explicit order get listed before sessions that don't have one.

This mechanism has the immense advantage of being simple to implement and of
not requiring any additional change to the project's setup. Among drawbacks:
- No checks are run on the specified orders for now: two sessions may
have the same expicit order (in which case they get sorted by issue number).
- There is no way to visualize the order directly or through a project's view.
- Edition is through the Note field, which may contain other text
  (e.g., '-warning:capacity')

Experience will tell whether we need a better mechanism for this...