Grid: fix shuffling and handling of CLI params
In the absence of a seed, the program shuffled the list randomly, which is
good, and generated a seed *afterwards* which was problematic because, by
definition, that seed wasn't used to shuffle the list!

The code no longer allowed "false" to be passed as a valid value for the
"apply" parameter, and failed to process the first parameter as a list of
issue numbers.