Validate the timezone defined in the project's metadata
The timezone needs to be one of the values that the W3C calendar system can
accept. In practice, that means one of the values in the official Time Zone
Database:
  https://www.iana.org/time-zones

... which should match values in the "tz Identifier" column in the following
Wikipedia page:
  https://en.wikipedia.org/wiki/List_of_tz_database_time_zones

Validating allows to trap the error early on, instead of ending up with some
hard to debug error in the code that updates the calendar entry.

This fixes #89.