Retrieve env variables from GitHub through "gh" CLI
There is no GraphQL endpoint to retrieve variables but the REST API can be used
in practice, and that's precisely what the `gh` CLI does. Using it to retrieve
variables means it's no longer needed to repeat variables locally in a
`config.json` file. This is useful for local execution.

Note: Secrets are still needed! They cannot be retrieved.