Add support for breakout events that span multiple days
This update makes it possible to organize a breakouts event that spans multiple
days. Sessions now need a room, a day and a slot within that day. For that to
work, the project must have an additional "Day" single select custom field that
lists possible days, using the format `YYYY-MM-DD` or `[name] (YYYY-MM-DD)`.
For a single day event, this field will have only one option.

The grid suggestion script was adjusted accordingly to consider all possible
days and slots, instead of just slots. Validation logic was also updated.

The code expects all days to be "equal", and in particular to have the exact
same slots. Additional constraints such as unavailable slots on a particular
day are not yet supported (a similar constraint that may be worth supporting
is "this room is not available on day 2 afternoon").

Important: sessions can still only be associated with one and only one room,
day and slot. There is no way to associate sessions with more than one slots
for now!