Add test infrastructure
The code creates a test mode where (most) network requests get replaced by a
stub function that returns test data from the `test/data` repository.

The update contains only one simple test for now, that basically makes sure
that the stubs work as intended.

Note that there is no way to keep a fully clean separation between code and
tests because there is no way to stub an ES6 module function in practice, due
to it being immutable when it's imported (for instance, `sinon.js` cannot be
used). The code continues to rely on environment variables to determine what
it needs to do.

GraphQL mutations requests will be prevented by the stub functions but they do
not yet return the expected answer.

Also, the test mode cannot be used to test the calendar code.