Add mechanism to reset internal cache states between tests
The code uses a number of internal memory caches, which is totally fine for
normal usage as the data is not supposed to change. From a testing perspective,
data will change though. This update creates suitable functions to reset the
state of the different caches between tests, and a generic `initTestEnv`
function that should be called before each test suite (or each time a test
needs to use different data).

The update also relaxes the check on the slot duration to allow slots whose
duration is anywhere between 30mn and 120mn (because group meetings slots may
last two hours).