Group meetings: Associate issues with groups instead of chairs
The issue title is supposed to be the name of the group. It must contain the
type of group (for instance "Second Screen WG"). The type may be in expanded
form (as in "Second Screen Working Group").

The code retrieves the list of possible groups from the W3C API. The code can
handle additional groups through the `GROUPS_W3CID` env variable if needed.

A few exceptions are hardcoded (TAG, CSS, WICG, PING) to allow short forms to
be used (e.g., "CSS WG" or "TAG") as well as to complete the label that will
be used to normalize the issue title ("CSS Working Group" would become
"Cascading Style Sheets WG (CSS)").

Calendar management logic now selects the right type of meeting (group meeting
or breakout), and either lists groups as attendees for group meetings or chairs
for breakouts.