Support multiple meetings for a single session (validation only)
This adds the possibility to define a new `Meeting` custom field in projects
that allows meeting organizers to schedule sessions more than once.

The `Meeting` field takes a list of tuples (room, day, slot). When the
`Meeting` field exists, the `room`, `day` and `slot` field no longer need to
be set. If they are set and there is no `Meeting`, they define a first meeting
for the session. If they are set while the `Meeting` value is set, they define
default values for things not further specified in the `Meeting` field.

tuples must be separated by ";" or "|". Parts of each tuple must be separated
by ",". White spaces are not significant. For example:
  "Room 1, Monday, 9:00; Room 1,Monday,11:00"

The room, day and slot identifiers may be the full identifiers, or a logical
part of them (the room name without its capacity, the day name, or the start
time of a slot).