Support multiple meetings for a single session (calendar)
This adds and checks the logic to combine contiguous slots into logical
calendar entries, and adds a function that suggests calendar update actions to
synchronize things.

The calendar code now goes through these update actions and updates calendar
entries accordingly. It then reports the calendar entries to the session
description, with a more complete label than we used to have, something like:
`- [day start end](url)`. This allows the code to compute the update actions
without having to load calendar pages one by one. Obviously, that only works
provided that no one tampers with the calendar entries directly and with that
part of the session descriptions!