Visualize project as an HTML page
This update moves the HTML page generation part of the `suggest-grid` script to
a dedicated module that can be used and tested in isolation. Main updates made
to the page that gets created:

1. It uses the project's "meeting" label as title
2. It starts with a quick overview of sessions that reports on the number of
errors, warnings, and yet-to-be-scheduled sessions
3. It contains one table per event day
4. The page is sectioned in proper sections with IDs
5. New styles derived from Pure CSS are used to highlight problems and style
tables.
6. The page lists things that still need to be addressed.

A new `view-grid` tool gets added to generate the HTML page from the command
line. Run with `npx view-grid > grid.html` (to save the result to a `grid.html`
file).

Note the page does not report on all validation errors (apart from in the first
overview table).