Meetings changes code now expects a YAML file
The changes file that could be sent to suggest-grid was a simple text file,
but the introduction of the `meeting` makes the logic slightly more complex.
The script now expects that changes file to be a YAML structure that looks
like:

```yaml
18:
  reset:
    - day
    - slot
  day: 2024-03-12
  room: Room 5
  meeting:
    - Monday
    - Tuesday
23:
  reset: all
  room: Room 1
42:
  reset: day
```

The code is somewhat flexible in what it accepts (though not *that* flexible!)

A couple of simple tests were added to check the meetings changes logic.