Allow unknown group meetings through CHAIR_W3CID variable
It is possible to create a breakout session with a chair that does not have a
W3C account by setting an entry in the `CHAIR_W3CID` variable with a `-1`
value, e.g., `{ 'tidoust': -1 }`.

This update makes the code reuse the variable for group meetings as well, so
that it becomes possible to authorize a meeting of a group that is not a real
W3C group, such as "WHATWG", "Web Platform Tests", "W3C Chapters", etc.

The code is also more flexible in what it accepts as value to note that the
group does not exist. Anything that is not a number will be interpreted to
mean that the group/chair is legit.

Note that we may want to rename the variable later on to something more
neutral than "chair".