Drop `upload-grid`, return grid data as YAML
The `upload-grid` tool has essentially become useless now that `suggest-grid`
handles everything. As such, the JSON data at the end of the grid page in HTML
was correct but useless. The HTML page now includes the data as YAML. The YAML
can directly be fed as a list of changes to the `suggest-grid` tool (with or
without adjustments) to re-create the schedule in a more predictable way than
through passing the seed (passing the seed only re-creates the same schedule if
data has not changed in the meantime).