Create one CLI to (almost) rule them all
This update replaces the `suggest-grid`, `validate-session`,
`validate-grid`, `update-calendar` and `view-events` tools with a single
`tpac-breakouts` CLI that comes with cleaner option handling and built-in
usage help.

To run and get help: `npx tpac-breakouts --help`.

Previous tools were deleted. Jobs that used them obviously need to be update
to target the new commands.

The update also fixed a couple of bugs in these tools. Still no tests for now
but the commands should be easier to fix now that they exist in a separate
`commands` folder and start from a project object.