Support "None" value for dropdown fields
For some reason, in private repositories, GitHub does not support "required"
fields in issue templates, see:
https://github.blog/changelog/2023-01-12-github-issues-january-12th-update/#%F0%9F%93%9D-issue-forms-for-private-repositories-public-beta

In practice, a "None" option gets added to required dropdown fields in private
repositories, which the code did not expect. The code now sets the value to a
default value when it bumps into "None": for the type of session, default is
"breakout". For capacity, default is 0.