Add mechanism to specify custom slot start/end times
See #100. If the start/end times of a meeting needs to be adjusted for some
reason, the actual start/end times may now be specified in the `Meeting` field
using `<HH:mm>` next to the time that needs to be modified.

For example:
- `9:00<8:30>`
- `9:00-11:00<10:30>`
- `11:00<11:15> - 13:00<12:45>`

Validation will reject problematic custom times:
- a custom start time that overlaps with a previous slot
- a custom end time that overlaps with a next slot
- custom times that do not change anything
- custom times that make no sense (e.g., because custom start time is later
than the custom end time)

The custom start/end times are used in the calendar.

If a group meets during contiguous slots and custom start/end times are set
for the inner slots, the custom start/end times are ignored. For example, the
custom times are ignored in `9:00; 11:00<11:15> - 13:00<12:45>; 14:00` because
this will map to only one calendar entry from `9:00` to `16:00`.