Add "Try me out" field mechanism to try schedule updates (#120)
This adds a new `try-changes` command that tries schedule updates defined in
a "Try me out" field, as discussed in #115.

The optional `--apply` flag tells the command to apply the adjusted schedule.
Applying the adjusted schedule resets the "Try me out" fields.

The command is restricted to group meetings for now. Running it on projects that
do not define a "Meeting" field will not work.