Support non W3C group names in joint meetings too
The code expected joint meetings would only involve known W3C groups, but there
may be joint meetings with other groups in practice (e.g., "WHATWG & CSS WG").

This update is a rewrite of the function that converts an issue title to a list
of groups. The function now consumes the title in chunks.

(This should also fix the "CSS" case, making use of the full name, name without
"(CSS)", or simply "CSS" possible)