Add groups view to generated HTML page (#127)
The grid view is great to look at the schedule. Groups may be more interested
in a short summary for themselves. This update adds a groups view next to the
grid in the generated HTML page.

The groups view also lists all warnings and errors that got raised against the
sessions related to the underlying group.