Do not schedule meetings that need a bigger room first
Scheduling meetings that need a bigger room first does not seem to yield
great results, because we can end up scheduling meetings that need a large
number of slots last with that approach, and slots may then no longer be
available for them.

This update drops the rule. Meetings that need a bigger number of slots are
scheduled first. Other constraints do not influence the order otherwise.