Support the concept of "VIP" rooms (#131)
Via #128. Some rooms may be dedicated to specific meetings, especially at TPAC.
To implement the functionality, rooms may now be defined with an addition
"(vip)" flag, as in: "Business room (75) (vip)" or "Plenary (VIP)". Casing does
not matter.

When the "(vip)" flag is set to a room's name, the scheduler no longer assigns
sessions to that room. A session may still explicitly be tied to a VIP room
through the `Room` field.

This mechanism allows to have a room dedicated to the Advisory Committee
meeting.