Fix handling of quotes at the end of lines
Double quotes were correctly escaped when preparing the GraphQL mutation query.
However, GraphQL has weird rules when it processes multi-line strings: things
work nicely even when the string is wrapped in a single double-quotes (instead
of in triple-quotes, which would be more appropriate for multi-line strings,
see: https://spec.graphql.org/draft/#sec-String-Value), but parsing does not
like it when it bumps onto a double-quotes at the end of a line, even if that
double-quotes is properly escaped.

The fix is to also escape new line characters to `\n`, so that the string
remains single-line (an alternative would be to wrap the string in
triple-quotes but that seems to change parsing rules slightly as well)

Fixes #139.