Support group meetings with highlights (#138)
An issue title may now contain an highlight, defined as any string after a `:`
or a `>` (and after the group name(s)). See #137 for details.

The full title (with highlight) is used in the grid and in the calendar.
Validation will not report an error when two session issues exist for the same
group when highlights differ. The scheduler will see issues as completely
separate. Note that it won't try to schedule the sessions in the same room as
a result.