Use `select` method to select big meeting value
The script that shows/hides the event category and restricted boxes hooks
onto the `change` event of the select field. The way we were selecting the
option did not trigger that event, and made the rest of the logic fail as it
couldn't find fields it was expecting.

The code now uses Puppeteer's `select` method, which triggers the `change`
event.