CLI: First stab at "assess-rooms" command
This adds an initial `assess-rooms` command to analyze rooms capacities
against known list of registrants for TPAC events, as described in #150.

To get a quick report in markdown, run:

```
npx tpac-breakouts assess-rooms all > res.md
```

The command only looks at the overall number of participants and observers for
each meeting. It does not expect entries in the registrants page to contain
duplicated entries (although the one for TPAC 2024 does for the APA WG for some
reason).

Report also remains simple for the time being.