Add new "switch" warning to alert on room changes
The `warning: switch` warning gets reported when a session needs to switch
rooms from one slot to the next. The warning is *not reported* if the room
change occurs from one day to the next.

In the HTML schedule page, the warning appears both in the schedule table
through a "Previous slot in: [room name]" message and in the "Groups View"
section.