Handle expanded group names in sessions with highlights
The code incorrectly took for granted that a session with an highlight would
always use acronyms for the types of groups (e.g., "WG" instead of "Working
Group"), and failed to recognize the name of the group if the expanded form
was used.

While it may be a good idea to keep session titles short, code should still
handle the expanded form. Done in this update.