Add ability to read room metadata from the room description
Implements part of https://github.com/w3c/tpac-breakouts/issues/134#issuecomment-2155995592,
making it possible to define the room's metadata within the description field
next to the definition of the room in the GitHub project.

Note that this does not make any practical change to available semantics. The
code does not yet support any more metadata key (such as "availability"). To
be done later on.