Add new "view-registrants" CLI command, replaces "assess-room-capacity"
The new command replaces the "assess-room-capacity" command and lets one
visualize the number of participants and observers of sessions, and generate a
report with sessions in large enough rooms, in rooms that are too small due to
the number of participants, and in rooms that are too small due to the number
of observers.

The numbers of registrants are retrieved from a `Registrants` custom field in
the project.

Command options make it possible to:
1. Fetch updated numbers from a registrants page (for TPAC events).
2. Update these numbers in the project's `Registrants` custom field.
3. Restrict the output to warnings (rooms too small)