list-chairs: "js" option to create a script for registrants page (#172)
Passing `js` as second argument to list-chairs will make it spit out a
JavaScript script that contains the list of chairs but also a function to
check the list against a page that lists registrants. The script needs to
be a copied and pasted into the dev console linked to the registrants page.
It returns the list of session chairs that still need to register.