Add Google spreadsheet AppScript code and bundling logic
This adds the Google AppScript code used to synchronize meetings data in a
GitHub repository with a Google spreadsheet.

The AppScript runtime does not (yet?) support JavaScript modules, so the
code needs to be bundled (done with Rollup) before it may be deployed with
Google's `clasp` CLI.

For deployment to work, user must have logged in into clasp with:

```
npx clasp login
```

Give the application enough permissions to create/update/delete script
projects.

The user must also create a `.clasp.json` file at the root of the project that
contains two entries:
- `scriptId`: the ID of the standalone Google AppScript that should contain the
resulting code. You must have permissions to update that script.
- `rootDir`: full path to the `tools/appscript/bundle.js` file (that file will
be generated from the source code in `tools/appscript`).

Once everything is setup, run the following command to update the script in the
Google Drive:

```
npm run appscript
```