minutes-to-w3c: convert Google Docs to HTML
The script converted Google Docs to PDF using Puppeteer. But Google Docs can
be exported to self-contained HTML pages instead (and Puppeteer is not needed
for that). This update downloads HTML versions of minutes taken in Google Docs
so that they may imported to W3C's web sites.

The script also returns the Markdown updates that need to be made to the issues
to link to the new minutes (and keep a pointer to the initial ones).

Notes:
- The script is not (yet) smart enough to udpate the issues itself.
- The script takes a second parameter: the prefix to use for the new link
on W3C's web site, e.g., `https://www.w3.org/2024/09/breakouts`.