New "import from GitHub" menu in a Google spreadsheet
The import no longer kicks off a GitHub workflow but rather pulls the data from
GitHub directly and refreshes the content of the spreadsheet.

Custom data associated with rooms, days, slots, or even breakouts in additional
columns is preserved. Formatting is preserved too. The breakouts sheet is
created if needed.