Save validation results in sheet and refresh the grid
Validation results are now recorded in the sessions sheet in the appropriate
Error, Warning, and Check columns.

Also, validating the grid requires re-generating the grid afterwards to render
the results. The logic was factored out as a result (but the logic now needs to
be updated to also report validation results)