Add "create repository" custom menu option and logic
The option runs a series of GraphQL queries to create and initialize a GitHub
repository for the event. Due to limitations in the GraphQL API, the logic needs
to initialize the repository from a (relatively empty) template repository,
currently sitting at https://github.com/tidoust/tpac-breakouts-template.

The repository is created as a private repository.