Always use latest version of tpac-breakouts in jobs
That means no need to bump the version of tools whenever an update is made to
the tools. Unfortunately, it also means there's no way to pin a repository to
a specific version of the tools, so older repositories are likely going to
break over time (but then they should essentially freeze after the event).