Drop dependency on GitHub project (#227)
The logic had already been partly re-written to rely on GitHub variables instead of on the presence of a GitHub Project. This update further adjusts the logic to use the variables throughout. As a result, a GitHub project no longer needs to be created when a new event gets created, only a repository is.

A new "Create repository" menu item got added to the "Advanced menu" of the Google spreadsheet. It takes care of what used to be done by the `create` CLI command (which no longer exists) and creates and initializes a GitHub repository for the event in particular. The menu item does not set GitHub secrets though (`GRAPHQL_TOKEN` and `W3C_PASSWORD`). They have to be set manually after creation (along with permissions for tpac-breakouts-bot).