Fix logic for TPAC group meetings (#228)
Recent updates were tailored at breakouts events and had broken the logic for
TPAC group meetings. This attempts to unbreak things.

With TPAC group meetings, a new "Meetings" sheet gets created to list
individual meetings for each group. Right now, columns in that sheet are
"number" (for the group issue number), "room", "day" and "slot". These columns
should get improved over time to make life easier for admins. For example, it
seems pretty useful to replace the "number" column with a "number - title"
column and to add data validation logic so that admins can spot meetings per
group. Also, the "day" column should probably contain the weekday instead of
the date to ease reading.

The "Meetings" sheet takes precedence over the "meeting" column in the
"Sessions" sheet. It would seem a good idea to get rid of the "meeting" column
altogether: we needed it when we were using a project, but we can now easily
store more complex structures in GitHub variables.

The update also fixes a number of bugs identified along the way, including code
that was still expecting a GitHub project, and the repo creation code that
failed to save the issue template within the spreadsheet, making it impossible
to get a schedule afterwards.

Schedule sheets now get created as read-only. One drawback is that this prompts
the user with a warning when they try to delete the sheet, which isn't
fantastic from a UX perspective (if users need to dismiss these warnings,
they'll be tempted to dismiss similar warnings that should *not* be dismissed).