Drop "Meeting" column in List view
Group meetings are directly handled through the "Meetings" sheet. Reporting
them in a "Meeting" column in the list view means the information appears in
two different places and creates a point of synchronization failure should
event admins decide to update meetings directly in the list view. This update
drops the column from the list view and makes sure that the info from the
"Meetings" sheet gets used when the schedule gets published.