Handle acceptable list of slots in scheduler
The scheduler now leverages the list of acceptable slots to propose a schedule.

If there are more acceptable slots than the requested number of slots, the
scheduler will favor slots that are early in the week, and will not favor
continuous slots. That's not ideal, but improving on this behavior is non
trivial.