Support links for the IRC channel
The code now accepts that the IRC channel be specified as a link and
automatically serializes the IRC channel as a link to webirc.w3.org.

Note the code does not attempt to understand or reuse the URL that got provided
in the initial link. It just considers that the IRC channel is the link label.

Fixes #177.