Merge days and slots (#235)
Days and slots used to be specified separately. That was done for TPAC events
with the assumption that slots would be the same for each day. That assumption
no longer holds true: next TPAC will use different slots per day. To get back
the flexibility we need, this update merges the concept of days and slots into
one flat list.

In the spreadsheet, that means one "Slots" sheet that ought to contain days
and slots (repeating days as needed), instead of two separate sheets.

Internally, the distinction between days and slots is preserved at the session
level because:
1. We may want to set the day(s) when a group meets by hand but let the
scheduler choose times.
2. We don't want to warn when a group needs to switch rooms across days.

The update also simplifies the mapping to the slot time to only use the start
time.