Combine day and slot in Meetings sheet
Because slot times may vary by day, presenting day and slots in different
columns in the spreadsheet is not fantastic. The Meetings sheet now only
features a "Slot" column that contains the concatenation of the day and start
time.

Note: It would make sense to display the weekday (Monday, Thursday, etc.)
instead of a pure date for human readability. To be addressed when the sheet
gets improved to also display the group's title.

Close #236