Stop using issue labels to record tracks
With this update, mapping between a session or group meeting and a track uses
a dedicated "Tracks" column in the "List" sheet. Issue labels are no longer
used.

This allows to specify tracks, which are an admin thing, within the spreadsheet
directly instead of having to go through GitHub.

This also allows to keep tracks as an internal thing only.

A new "Show track information in calendar" setting can be set in the "Event"
sheet to say whether track information should appear in calendar entries. It's
probably a good idea to keep it to "no", but since the code already exists to
fill out that part, I thought I'd keep it around in case we want to reverse
our resolution not to expose tracks anymore.