Drop all logic related to rooms labels
Rooms used to get their name through a GitHub project and the room name used to
include indications about the room's capacity and on whether the room is a VIP
room. The code computed a room's label without metadata on top of the room's
name, and used the label whenever the info was targeted at end users.

While it was already possible to specify additional metadata using the room's
description in the GitHub project, with the switch to a spreadsheet, completing
the room name with metadata is simply meaningless.

There was still some logic that accepted labels... and that also created bugs
because some bits actually *required* labels to be present, while they were now
only really generated in tests.

This update drops labels throughout, and adjusts tests accordingly.