Get rid of `day` for sessions throughout (#250)
The `slot` property of sessions stores both the date and the start time of the
meeting. It used to store only the meeting times, while the date used to be
captured in a `day` property.

The code still handled or incorrectly *expected* the day property here and
there. This update properly drops the possibility to set a schedule day at the
session level in a `day` property, and adjusts tests accordingly.

Internally, atomic meetings are presented with a `day` property that contains
the date of the meeting and a `slot` property that contains the start time.
That's normal although certainly unfortunate that the same terminology be used.
Merged meetings rather use `start` to record the meeting start time.