Validate sessions after import
Refreshing sessions used to retrieve the validation messages from GitHub, which
may already be outdated (for example, when an unknown group got added to the
list, or simply when scheduling data in the spreadsheet is different from that
in the GitHub repository).

The refresh menu now automatically re-validates the list of sessions and
updates the Error, Warning, Check columns in the List view accordingly.