Don't serialize session desc if a format error was found
A recent change made the job that validates a session re-serialize the
description every time to make sure that the "Avoid conflicts" section gets
added to all sessions. But serialization of the description can only succeed if
session follows the right format.

The job validated sessions with format issues but crashed in the end since it
couldn't serialize the description again. This update makes it only attempt
that when no formatting error was detected.