Avoid conflicts with joint meetings of conflicting groups too
Via #255. When Group A says "don't conflict with group Z" then the scheduler
should also avoid conflicts between the A schedule and any (joint) meeting
involving Z.

This update applies the logic to the scheduler and validation. It also
completes the description of group sessions in the schedule grid with the list
of indirect conflicts, highlighting those that create problems.