Fetch registration information from W3C endpoint
This makes the code fetch registration information from a dedicated endpoint.
For that to be possible, the code needs:
1. the big meeting's slug. To be added to the Event sheet. Longer term, this
should completely replace the big meeting's name. In the meantime, the name is
still needed to manage interaction with the calendar.
2. an authorization token. To be set once through a dedicated "Advanced" menu.
The token is stored as developer metadata within the spreadsheet. This at least
hides it from sight, as tokens should be (any sufficiently motivated person
would be able to retrieve it, but then the spreadsheet should remain team-only
so it does not really matter).

With these two pieces of information, the code retrieves the information from a
W3C server:
- Registration information gets stored in the Meetings sheet
- Chairs/Staff contact information gets stored in a People sheet

Validation uses that information to alert about capacity issues.

Note: The scheduler does not look at the number of participants when it assigns
rooms for now.