Add export view that mimics info in W3C admin interface
To ask TPAC registrants about what groups they plan to attend, the list of
group meetings needs to be added to the registration form in the W3C admin
interface.

This process could be automated in the long term. In the short term, this
proposes a new HTML file to the "Export event to files" advanced menu that
mimics what the resulting table should look like in the W3C admin interface:
https://www.w3.org/admin/tpac-meetings/list