Take registered participants into account in scheduler
Close #268. This adds support for assigning rooms based on the actual
number of registered participants in the scheduler.

The scheduler does not look at nuanced registration data per day for each
session. It rather assumes that it always needs to find a room that can
accommodate the maximum number of participants for the session across all days.

This should work well because we try to schedule groups to meet in one and only
one room in any case.